refactor(service): fake 객체를 service.fake 패키지로 이동
